@@ -3,7 +3,7 @@ export default interface ICreateContactRequest {
   name: string;
   email: string;
   phone: string;
-  category_id: string;
+  categoryId: string;
 }
 
 export type IUpdateContactRequest = ICreateContactRequest
