@@ -1,3 +1,6 @@
+import ICreateContactRequest from '../HTTP/requests/IContactRequest';
+import IResponseContactRequest from '../HTTP/responses/IContactResponse';
+import ContactMapper, { DomainContactResponse } from './mappers/ContactMapper';
 import HttpClient from './utils/HttpClient';
 
 class ContactsService {
@@ -7,24 +10,32 @@ class ContactsService {
     this.httpClient = new HttpClient(process.env.REACT_APP_BASE_URL || 'http://localhost:3001');
   }
 
-  listContacts<T>(orderBy = 'asc'): Promise<T> {
-    return this.httpClient.get<T>(`/contacts?orderBy=${orderBy}`);
+  async listContacts(orderBy = 'asc'): Promise<DomainContactResponse[]> {
+    const contacts = await this.httpClient.get<IResponseContactRequest[]>(`/contacts?orderBy=${orderBy}`);
+
+    return contacts.map(ContactMapper.toDomain);
   }
 
-  createContact<D, R>(formData: D): Promise<unknown> {
-    return this.httpClient.post<D, R>('/contacts', {
-      body: formData,
+  createContact<R>(contact: ICreateContactRequest): Promise<unknown> {
+    const body = ContactMapper.toPersistence(contact);
+
+    return this.httpClient.post<R>('/contacts', {
+      body,
     });
   }
 
-  updateContact<D, R>(id: string, formData: D): Promise<R> {
-    return this.httpClient.put<D, R>(`/contacts/${id}`, {
-      body: formData,
+  updateContact<R>(id: string, contact: ICreateContactRequest): Promise<R> {
+    const body = ContactMapper.toPersistence(contact);
+
+    return this.httpClient.put<R>(`/contacts/${id}`, {
+      body,
     });
   }
 
-  getContactById<T>(id: string): Promise<T> {
-    return this.httpClient.get<T>(`/contacts/${id}`);
+  async getContactById(id: string): Promise<DomainContactResponse> {
+    const contact = await this.httpClient.get<IResponseContactRequest>(`/contacts/${id}`);
+
+    return ContactMapper.toDomain(contact);
   }
 
   deleteContact<R>(id: string): Promise<R> {
